app.js separated - constants.js and utils.js created
